added module in script
The error you would see in the console if you tried to use an import statement outside a module in JavaScript would typically be something like:

javascript
Copy code
Uncaught SyntaxError: Cannot use import statement outside a module
This error occurs because import statements can only be used inside a module, which is a file that has been explicitly marked as a module using the type="module" attribute on the <script> tag or by using the .mjs file extension. When you try to use an import statement outside of a module, the JavaScript interpreter does not recognize it as valid syntax and throws an error.