Update index.html
t's possible that the statements in your script tag are being output to the page because they are not enclosed within an HTML element.

When you include a script tag within the body of an HTML document, any statements within the script tag will be executed by the browser when it reaches that part of the document. If you want to output any content to the page, you need to enclose it within an HTML element, such as a div, span, or p tag.

For example, you could modify your code like this to output a message to the page: